Drop unnecessary verification code cycling
After thinking about this a bit more, the only thing that needs to
happen is the frontend performs an async resend of the email.

Because verification codes don't expire, there's no reason to cycle
them.  The only advantange of expiring codes is to prevent
automated signup using brute force verification.  That shouldn't be
a big issue, because brute forcing a uuid4 would require enough load
that alarms/monitoring should catch it.

The alternative is extra storage (for expiry) and a lot more code
to rotate verifications, and send email from any touch point that
realizes the code has expired.